proxy-tester/requirements.txt
cassandra-driver==3.29.2
requests==2.32.3
PySocks==1.7.1
